+++ main-server/src/core/as/process_data_for_rp.js
@@ -256,6 +256,7 @@ async function processDataForRPInternalAsync(
       );
       const signatureBuffer = await utils.createSignature(
         signingPublicKey.algorithm,
+        signingPublicKey.version,
         data + data_salt,
         nodeId
       );