+++ main-server/src/env_var_validate.js
@@ -153,7 +153,14 @@ if (process.env.DEFAULT_API_VERSION != null) {
   }
 }
 
-const CALLBACK_API_VERSION_ALLOWED_VALUES = ['4', '4.0', '5', '5.2', '6', '6.0'];
+const CALLBACK_API_VERSION_ALLOWED_VALUES = [
+  '4',
+  '4.0',
+  '5',
+  '5.2',
+  '6',
+  '6.0',
+];
 if (process.env.CALLBACK_API_VERSION != null) {
   if (
     !CALLBACK_API_VERSION_ALLOWED_VALUES.includes(
@@ -163,6 +170,26 @@ if (process.env.CALLBACK_API_VERSION != null) {
     console.error(
       'ERROR:',
       'Unsupported "CALLBACK_API_VERSION" environment variable value. Only "4.0", "5.2", and "6.0" are allowed. Process will now exit.'
+    );
+    process.exit(1);
+  }
+}
+
+const EXTERNAL_CRYPTO_SERVICE_CALLBACK_API_VERSION_ALLOWED_VALUES = [
+  '5',
+  '5.2',
+  '6',
+  '6.0',
+];
+if (process.env.EXTERNAL_CRYPTO_SERVICE_CALLBACK_API_VERSION != null) {
+  if (
+    !EXTERNAL_CRYPTO_SERVICE_CALLBACK_API_VERSION_ALLOWED_VALUES.includes(
+      process.env.EXTERNAL_CRYPTO_SERVICE_CALLBACK_API_VERSION
+    )
+  ) {
+    console.error(
+      'ERROR:',
+      'Unsupported "EXTERNAL_CRYPTO_SERVICE_CALLBACK_API_VERSION" environment variable value. Only "5.2", and "6.0" are allowed. Process will now exit.'
     );
     process.exit(1);
   }